[SES-52] Created MVVM for transparency reporting
